allow symlink as input
Previous implementation used 'lstatSync'  which returns stat info about link
itself, not the file that it refers to. That behaviour didn't allow to use the
tool in case of separate build (source and build dirs are separate, and sources
linked back into  build directory).

As workaround there we could provide file content through stdin, but that
approach has limitations:
- you cannot process multiple files at once
- you cannot use 'include' directives

This commit changes the use of  'lstatSync' to 'statSync' which makes symlink
files be processed as expected.